feat(backend): add native-wayland delayed clipboard data source
Split linux_clipboard into a module: shared pure formatting/fetch logic
plus x11 and wayland delayed-render owners. The wayland path owns a
wlr-data-control data source and serves text/uri-list and gnome/mate
copied-files on the send callback, fetching remote files only on paste.

Refs #1847
